--- src/HomeTask1/Triangle/Point.java
@@ -1,20 +1,0 @@
-package HomeTask1.Triangle;
-
-public class Point {
-    double x;
-    double y;
-
-    public Point(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Point point = (Point) o;
-        return Double.compare(point.x, x) == 0 &&
-                Double.compare(point.y, y) == 0;
-    }
-}
