--- src/HomeTask1/Triangle/TriangleException.java
@@ -1,5 +1,0 @@
-package HomeTask1.Triangle;
-
-
-public class TriangleException extends Exception {
-}
