--- src/HomeTask1/aaa.java
@@ -1,4 +1,0 @@
-package HomeTask1;
-
-public class aaa {
-}
