--- src/HomeTask2/Area/CircleArea.java
@@ -1,7 +1,0 @@
-package HomeTask2.Area;
-
-public class CircleArea extends Shape {
-    public double getArea() {
-        return Math.PI * (getValue() * getValue());
-    }
-}
