--- src/HomeTask2/Area/RectangleArea.java
@@ -1,7 +1,0 @@
-package HomeTask2.Area;
-
-public class RectangleArea extends Shape {
-    public double getArea() {
-        return getHeight() * getWidth();
-    }
-}
