--- src/HomeTask2/Area/SquareArea.java
@@ -1,8 +1,0 @@
-package HomeTask2.Area;
-
-public class SquareArea extends Shape {
-    @Override
-    public double getArea() {
-        return (getValue() * getValue());
-    }
-}
