--- src/HomeTask2/Bank/Observer.java
@@ -1,5 +1,0 @@
-package HomeTask2.Bank;
-
-public interface Observer {
-    void update(String var1);
-}
