--- src/HomeTask3/Task4/Passport.java
@@ -1,16 +1,0 @@
-package HomeTask3.Task4;
-
-public class Passport {
-    public String series;
-    public String number;
-
-    public Passport(String series, String number)
-    {
-        this.series = series;
-        this.number = number;
-    }
-
-    public String getFullId() {
-        return this.series + this.number;
-    }
-}
