+++ src/OOPHomeTask1/Triangle/TriangleException.java
@@ -1,0 +1,5 @@
+package OOPHomeTask1.Triangle;
+
+
+public class TriangleException extends Exception {
+}
