+++ src/OOPHomeTask1/aaa.java
@@ -1,0 +1,4 @@
+package OOPHomeTask1;
+
+public class aaa {
+}
