+++ src/OOPHomeTask2/Area/CircleArea.java
@@ -1,0 +1,7 @@
+package OOPHomeTask2.Area;
+
+public class CircleArea extends Shape {
+    public double getArea() {
+        return Math.PI * (getValue() * getValue());
+    }
+}
