+++ src/OOPHomeTask2/Area/RectangleArea.java
@@ -1,0 +1,7 @@
+package OOPHomeTask2.Area;
+
+public class RectangleArea extends Shape {
+    public double getArea() {
+        return getHeight() * getWidth();
+    }
+}
