+++ src/OOPHomeTask2/Area/SquareArea.java
@@ -1,0 +1,8 @@
+package OOPHomeTask2.Area;
+
+public class SquareArea extends Shape {
+    @Override
+    public double getArea() {
+        return (getValue() * getValue());
+    }
+}
