+++ src/OOPHomeTask2/Bank/LogObserver.java
@@ -1,0 +1,10 @@
+package OOPHomeTask2.Bank;
+
+public class LogObserver implements Observer {
+    public LogObserver() {
+    }
+
+    public void update(String message) {
+        System.out.println("[INFO] => " + message);
+    }
+}
