+++ src/OOPHomeTask2/Bank/Observer.java
@@ -1,0 +1,5 @@
+package OOPHomeTask2.Bank;
+
+public interface Observer {
+    void update(String var1);
+}
