+++ src/OOPHomeTask3/Task4/Passport.java
@@ -1,0 +1,16 @@
+package OOPHomeTask3.Task4;
+
+public class Passport {
+    public String series;
+    public String number;
+
+    public Passport(String series, String number)
+    {
+        this.series = series;
+        this.number = number;
+    }
+
+    public String getFullId() {
+        return this.series + this.number;
+    }
+}
