+++ src/jUnit/TestSuit.java
@@ -1,0 +1,14 @@
+package jUnit;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+
+@RunWith(Suite.class)
+
+@Suite.SuiteClasses({
+        DecodingTest.class,
+        EncodingTest.class,
+        MyLinkedListTest.class
+})
+public class TestSuit {
+}
